Move code for snippets
