+++ docs/src/main/paradox/spring-web.md
@@ -30,14 +30,14 @@ framework, which means that Spring is made aware of Sources and Sinks etc.
 All you need to do is include the above dependency (`akka-stream-alpakka-spring-web`), start your app as usual:
 
 Java
-: @@snip [snip](/spring-web/src/test/java/akka/stream/alpakka/spring/web/DemoApplication.java) { #use }
+: @@snip [snip](/spring-web/src/test/java/docs/javadsl/DemoApplication.java) { #use }
 
 
 And you'll be able to return Akka Streams in HTTP endpoints directly:
 
 
 Java
-: @@snip [snip](/spring-web/src/test/java/akka/stream/alpakka/spring/web/SampleController.java) { #use }
+: @@snip [snip](/spring-web/src/test/java/docs/javadsl/SampleController.java) { #use }
 
 Both `javadsl` and `scaladsl` Akka Stream types are supported.
