+++ spring-web/src/test/java/docs/javadsl/DemoApplication.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.spring.web;
+package docs.javadsl;
 
 // #use
 import org.springframework.boot.SpringApplication;
