+++ spring-web/src/test/java/docs/javadsl/SampleController.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.spring.web;
+package docs.javadsl;
 
 // #use
 import akka.NotUsed;
